backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
cssutils==2.11.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
emails==0.6
fastapi==0.115.5
h11==0.14.0
idna==3.10
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
more-itertools==10.5.0
passlib==1.7.4
postgres==4.0
premailer==3.10.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyasn1==0.6.1
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
